Remove the backup code logic from the verify profile concern
The verify profile concern is intened to redirect users with profiles that need additional work to verify or unlock to the appropriate place. The backup code reminder logic got mixed into this which causes a little confusion and indirection. This commit moves that logic into its own concern.

[skip changelog]
